Add clang-tidy linter and fix all violations
Adds .clang-tidy config (bugprone/modernize/performance/readability groups
with noisy checks suppressed) and a lint CMake target. Fixes all existing
violations across the codebase: static_cast for narrowing conversions,
emplace_back, auto for cast-initialized vars, enum class, raw-string literals.

+++ src/app_state.cpp
@@ -1,7 +1,10 @@
 #include "app_state.h"
 
 static ImVec4 hex_to_imvec4(uint32_t rgb) {
-    return {((rgb >> 16) & 0xFF) / 255.0f, ((rgb >> 8) & 0xFF) / 255.0f, (rgb & 0xFF) / 255.0f, 1.0f};
+    return {static_cast<float>((rgb >> 16) & 0xFF) / 255.0f,
+            static_cast<float>((rgb >> 8) & 0xFF) / 255.0f,
+            static_cast<float>(rgb & 0xFF) / 255.0f,
+            1.0f};
 }
 
 PaletteState::PaletteState() {
